Confirmation leave team
Style fix

+++ resources/views/team/index.blade.php
@@ -11,7 +11,7 @@
                 @if($user->pivot->user_id == $userId)
                         <div class="card border-secondary mb-3" style="max-width: 18rem;">
                             <div class="card-header">{{ $team->name }}
-                                <a href="/delete/team/{{$team->id}}/{{$userId}}">x</a>
+                                <a href="/delete/team/{{$team->id}}/{{$userId}}" style="float: right" onclick="return confirm('Are you sure to leave the team?')">x</a>
                             </div>
                             <div class="card-body text-secondary">
                                 <a href="/team/{{$team->id}}" class="card-title">Go to {{ $team->name }}</a>
